fix: Use atomic as a context manager
Removed the `atomic` decorator of from the `AutoSingleRelatedObjectDescriptor.__get__` method and replaced it by using atomic as a context manager within the method when actually creating the new instance.

The previous behavior caused issues with other 3rd party packages.
Notably: https://github.com/strawberry-graphql/strawberry-django

Further context:
That field will contribute that descriptor to the class, which means `ModelClass.<attr>` will trigger that `__get__` which does not need to be atomic unless an instance will actually get created. This caused introspection issues with the optimizer in the above mentioned package.

Credit goes to @bellini666 for helping me find the source of the issue as well as providing the base patch. 🙏